move plotting functions to separate module
