+++ oasis/functions.py
@@ -1,7 +1,6 @@
 import numpy as np
 import scipy
 import scipy.signal
-import matplotlib.pyplot as plt
 from math import sqrt, log, exp
 from oasis import constrained_oasisAR1, oasisAR1
 from warnings import warn
@@ -13,26 +12,6 @@ except:
     cvxpy_installed = False
     warn("Could not find cvxpy. Don't worry, you can still use OASIS, " +
          "just not the slower interior point methods we compared to in the papers.")
-
-
-def init_fig():
-    """change some defaults for plotting"""
-    plt.rc('figure', facecolor='white', dpi=90, frameon=False)
-    plt.rc('font', size=30, **{'family': 'sans-serif', 'sans-serif': ['Computer Modern']})
-    plt.rc('lines', lw=2)
-    plt.rc('text', usetex=True)
-    plt.rc('legend', **{'fontsize': 24, 'frameon': False, 'labelspacing': .3, 'handletextpad': .3})
-    plt.rc('axes', linewidth=2)
-    plt.rc('xtick.major', size=10, width=1.5)
-    plt.rc('ytick.major', size=10, width=1.5)
-
-
-def simpleaxis(ax):
-    """plot only x and y axis, not a frame for subplot ax"""
-    ax.spines['top'].set_visible(False)
-    ax.spines['right'].set_visible(False)
-    ax.get_xaxis().tick_bottom()
-    ax.get_yaxis().tick_left()
 
 
 def gen_data(g=[.95], sn=.3, T=3000, framerate=30, firerate=.5, b=0, N=20, seed=13):
